noise/requirements.txt
# Dependencies for the CV container.

fastapi==0.115.12
uvicorn[standard]==0.34.2
pillow